# Conflicts:
#	Project/front/mymodules/Grad_cam.py


Former-commit-id: 04247af2ad85f2f3028f97a885cae5995baca480

+++ Project/front/mymodules/test_function2.py
@@ -101,7 +101,7 @@ def test(model_name, window):
             diagonal += cm[i][i]
             for j in cm[i]:
                 ssum += j
-        acc = round(diagonal / ssum, 3) * 100
+        acc = int(round(diagonal / ssum, 3) * 100)
 
         precision = []
         recall = []
